simple OOP-refactoring
Упрощён интерфейс класса. Нормировка весов колосьев в пересекающихся делянках и рассчёт их по участкам сокрыты в конструкторе класса.
